requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.9

openai==1.51.0
numpy==1.26.4
faiss-cpu==1.8.0.post1
pypdf==5.0.0
Pillow==10.4.0
pytesseract==0.3.13

python-dotenv==1.0.1

langchain-core
langchain-community




